requirements.txt
click>=6.7
Flask>=0.12.2
Flask-Admin>=1.5.0
Flask-SQLAlchemy>=2.2
flask-cache>=0.13.1
flask-login>=0.4.0
flask-wtf>=0.14.2
Logbook>=1.0.0
psycopg2>=2.7.1
pybase62>=0.3.1
temporaluuid64>=0.1.2
beautifulsoup4>=4.6.0
requests>=2.18.0
babel>=2.4.0
pytest>=3.1.2
boto3>=1.4.4
gunicorn>=19.7.1
alembic>=0.9.2
